requirements.txt
accelerate==0.27.2
beautifulsoup4==4.12.2
beir==2.0.0
codecarbon==1.2.0
cookiecutter==1.7.3
datasets==2.14.7
dspy-ai==2.4.5
evaluate==0.4.1
faiss-cpu==1.7.3
fschat[model_worker]>=0.2.35
ftfy>=6.1.3
fugashi==1.3.0
hf-doc-builder==0.4.0
hydra-core==1.3.2
hydra-submitit-launcher==1.1.5
ipadic==1.0.0
isort==5.12.0
kenlm==0.2.0
librosa==0.10.1
llama_index==0.9.3
loky==3.4.1
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.16.2
onnxruntime-tools==1.7.0
optuna==3.4.0
parameterized==0.9.0
peft==0.9.0
phonemizer==3.2.1
pyctcdecode==0.5.0
pyserini
pytesseract==0.3.10
pytest-timeout==2.2.0
pytest-xdist==3.4.0
python-dotenv>=1.0.1
rank-bm25==0.2.2
ray==2.8.0
requests-toolbelt==0.10.1
rhoknp==1.3.0
rjieba==0.1.11
rouge-score==0.1.2
ruff==0.0.259
sacrebleu==1.5.1
sacremoses==0.1.1
sentencepiece==0.1.99
SudachiDict-core==20230927
tensorboardX==2.6.2.2
timeout-decorator==0.5.0
torchaudio==2.1.0
tqdm==4.66.1
transformers>=4.35.2
unidic==1.1.0
unidic-lite==1.0.8
wandb==0.16.0
